Improvements to ASCII file load
Increased verbosity when loading ASCII files and storing either trues or scatter. As long as verbosity is set to true, will inform on the type of scatter components that have been selected (and will be saved) as well as the components excluded in trues. Trues will now always exclude randoms and Compton scattered events in the phantom, but the rest of the scatter events are excluded only if they are selected in the scatter components (no need to store scatter though).