Add support for 3D masks
This commit adds support for 3D masks in OpenCL and CUDA for MATLAB/Octave ONLY. This means that you can now input either the image-mask for backprojection or measurement-mask for forward projection as a 3D matrix. The size has to be the same as the image volume and the measurement data, respectively. 2D masks can also be used as before, but the 3D masks allow unique mask for each image slice or measurement projection/sinogram. Also includes a small fix for Nvidia-based GPUs in OpenCL for certain rare cases of large number of detector coordinates. This fix applies for both MATLAB/Octave and Python.