Add support for standalone regularizers
This commit adds support for standalone regularizer functions in Python. These can be used in custom reconstructions or completely independent of OMEGA. Simply input a vector of an image in PyOpenCL, ArrayFire, CuPy or Pytorch format and you can compute the gradient-based regularizer. Support is currently available for RDP, non-local regularizers and gradient-based TV.