Improve extended FOV and projection extrapolation
Both the extended FOV (with multi-resolution) and projection extrapolation now supports separate adjustable parameters for both the transaxial and axial components. This means that you can have varying extensions in either. Alternatively, you can still use the same global adjustment as before. Furthermore, the default behavior has changed such that extrapolation length is now 0.25 instead of 0.2 by default. For extended FOV, the default for transaxial remains the same (0.4), but axial is reduced to 0.3.